[GOTY-56] отрефакторил api, побил на модули
