+++ src/app/api/authApi.ts
@@ -1,0 +1,16 @@
+import { AxiosResponse } from 'axios'
+
+import { api } from './index'
+
+enum RequestRoot {
+  createToken = '/createToken',
+  invalidateToken = '/invalidateToken'
+}
+
+const baseURL = '/auth'
+
+export const createToken = (userLogin: string): Promise<AxiosResponse> =>
+  api.post(RequestRoot.createToken, { userLogin }, { baseURL })
+
+export const invalidateToken = (): Promise<AxiosResponse> =>
+  api.post(RequestRoot.invalidateToken, {}, { baseURL })
