+++ src/app/api/index.ts
@@ -1,0 +1,6 @@
+import axios from 'axios'
+
+export const api = axios.create({
+  withCredentials: true,
+  timeout: 10000
+})
